Divergesia/requirements.txt
PySide6>=6.0
Pillow>=9.5
numpy>=1.24
scikit-learn>=1.2
pytest>=7.3